1) created header file with declaring functions of date, arcane, etc 2) created C file with functions description of date, arcane, etc
