+++ etu_Akimov/etu_Akimov.vcxproj
@@ -127,12 +127,16 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="add_funcs.c" />
     <ClCompile Include="Arcane.c" />
     <ClCompile Include="test.c" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="info.txt" />
     <Text Include="log.txt" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="add_funcs.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
